+++ test/integration/nested_model.cc
@@ -324,8 +324,10 @@ TEST(NestedModel, NestedInclude)
   // each model has 3 links, and the link names of the nested models have
   // been transformed
   EXPECT_EQ(3u, model1->LinkCount());
-  EXPECT_EQ(3u, model2->LinkCount());
-  EXPECT_EQ(3u, model3->LinkCount());
+  // TODO (addisu): Update the following two expectations to account for the
+  // fact that included models are no longer flattened.
+  // EXPECT_EQ(3u, model2->LinkCount());
+  // EXPECT_EQ(3u, model3->LinkCount());
   const sdf::Link *baseLink1 = model1->LinkByName("base");
   const sdf::Link *baseLink2 = model2->LinkByName(name + "::base");
   const sdf::Link *baseLink3 = model3->LinkByName(name + "_14::base");
@@ -353,24 +355,14 @@ TEST(NestedModel, NestedInclude)
   EXPECT_EQ(lowerLink3->RawPose(), lowerLink1->RawPose());
   EXPECT_EQ(upperLink2->RawPose(), upperLink1->RawPose());
   EXPECT_EQ(upperLink3->RawPose(), upperLink1->RawPose());
-  // expect //pose/@relative_to to contain the name of the nested model frame
-  // injected during parsing. model1 is not nested, so it's links should have
-  // empty //pose/@relative_to
-  EXPECT_TRUE(baseLink1->PoseRelativeTo().empty());
-  EXPECT_EQ(name + "::__model__", baseLink2->PoseRelativeTo());
-  EXPECT_EQ(name + "_14::__model__", baseLink3->PoseRelativeTo());
-  EXPECT_TRUE(lowerLink1->PoseRelativeTo().empty());
-  EXPECT_EQ(name + "::__model__", lowerLink2->PoseRelativeTo());
-  EXPECT_EQ(name + "_14::__model__", lowerLink3->PoseRelativeTo());
-  EXPECT_TRUE(upperLink1->PoseRelativeTo().empty());
-  EXPECT_EQ(name + "::__model__", upperLink2->PoseRelativeTo());
-  EXPECT_EQ(name + "_14::__model__", upperLink3->PoseRelativeTo());
 
   // each model has 2 joints, and the joint names of the nested models have
   // been transformed
   EXPECT_EQ(2u, model1->JointCount());
-  EXPECT_EQ(2u, model2->JointCount());
-  EXPECT_EQ(2u, model3->JointCount());
+  // TODO (addisu): Update the following two expectations to account for the
+  // fact that included models are no longer flattened.
+  // EXPECT_EQ(2u, model2->JointCount());
+  // EXPECT_EQ(2u, model3->JointCount());
   auto *lowerJoint1 = model1->JointByName("lower_joint");
   auto *lowerJoint2 = model2->JointByName(name + "::lower_joint");
   auto *lowerJoint3 = model3->JointByName(name + "_14::lower_joint");
@@ -407,8 +399,6 @@ TEST(NestedModel, NestedInclude)
   // //axis/xyz/@expressed_in == "__model__" inside the nested model
   EXPECT_EQ(ignition::math::Vector3d::UnitX, lowerAxis3->Xyz());
   EXPECT_EQ(ignition::math::Vector3d::UnitX, upperAxis3->Xyz());
-  EXPECT_EQ(name + "_14::__model__", lowerAxis3->XyzExpressedIn());
-  EXPECT_EQ(name + "_14::__model__", upperAxis3->XyzExpressedIn());
 }
 
 //////////////////////////////////////////////////
@@ -452,6 +442,9 @@ TEST(NestedModel, NestedModelWithFrames)
   const sdf::Model *parentModel = world->ModelByIndex(0);
   ASSERT_NE(nullptr, parentModel);
 
+  const sdf::Model *childModel = parentModel->ModelByIndex(0);
+  ASSERT_NE(nullptr, childModel);
+
   using ignition::math::Pose3d;
   using ignition::math::Vector3d;
   // Expected poses for frames, links and joints after nesting.
@@ -463,55 +456,52 @@ TEST(NestedModel, NestedModelWithFrames)
   Vector3d joint1AxisExpVector = frame2ExpPose.Rot() * Vector3d::UnitZ;
   Vector3d joint1Axis2ExpVector = frame2ExpPose.Rot() * Vector3d::UnitX;
 
-  const auto *frame1 = parentModel->FrameByName("M1::F1");
-  ASSERT_NE(nullptr, frame1);
   Pose3d frame1Pose;
-  EXPECT_TRUE(frame1->SemanticPose().Resolve(frame1Pose).empty());
-  EXPECT_EQ(frame1ExpPose, frame1Pose);
+  auto parentSemPose = parentModel->SemanticPose();
+  EXPECT_TRUE(parentSemPose.Resolve(frame1Pose, "ParentModel::M1::F1").empty());
+  EXPECT_EQ(frame1ExpPose, frame1Pose.Inverse());
 
-  const auto *frame2 = parentModel->FrameByName("M1::F2");
-  ASSERT_NE(nullptr, frame2);
   Pose3d frame2Pose;
-  EXPECT_TRUE(frame2->SemanticPose().Resolve(frame2Pose).empty());
-  EXPECT_EQ(frame2ExpPose, frame2Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(frame2Pose, "ParentModel::M1::F2").empty());
+  EXPECT_EQ(frame2ExpPose, frame2Pose.Inverse());
 
   const auto *link1 = parentModel->LinkByName("M1::L1");
   ASSERT_NE(nullptr, link1);
   Pose3d link1Pose;
-  EXPECT_TRUE(link1->SemanticPose().Resolve(link1Pose).empty());
-  EXPECT_EQ(link1ExpPose, link1Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(link1Pose, "ParentModel::M1::L1").empty());
+  EXPECT_EQ(link1ExpPose, link1Pose.Inverse());
 
   const auto *visual1 = link1->VisualByName("V1");
   ASSERT_NE(nullptr, visual1);
-  EXPECT_EQ("M1::F2", visual1->PoseRelativeTo());
+  EXPECT_EQ("F2", visual1->PoseRelativeTo());
 
   const auto *collision1 = link1->CollisionByName("C1");
   ASSERT_NE(nullptr, collision1);
-  EXPECT_EQ("M1::__model__", collision1->PoseRelativeTo());
+  EXPECT_EQ("__model__", collision1->PoseRelativeTo());
 
   const auto *link2 = parentModel->LinkByName("M1::L2");
   ASSERT_NE(nullptr, link2);
   Pose3d link2Pose;
-  EXPECT_TRUE(link2->SemanticPose().Resolve(link2Pose).empty());
-  EXPECT_EQ(link2ExpPose, link2Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(link2Pose, "ParentModel::M1::L2").empty());
+  EXPECT_EQ(link2ExpPose, link2Pose.Inverse());
 
   const auto *joint1 = parentModel->JointByName("M1::J1");
   ASSERT_NE(nullptr, joint1);
   Pose3d joint1Pose;
-  EXPECT_TRUE(joint1->SemanticPose().Resolve(joint1Pose, "__model__").empty());
-  EXPECT_EQ(joint1ExpPose, joint1Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(joint1Pose, "ParentModel::M1::J1").empty());
+  EXPECT_EQ(joint1ExpPose, joint1Pose.Inverse());
 
   const auto joint1Axis = joint1->Axis(0);
   ASSERT_NE(nullptr, joint1Axis);
   Vector3d joint1AxisVector;
   EXPECT_TRUE(joint1Axis->ResolveXyz(joint1AxisVector, "__model__").empty());
-  EXPECT_EQ(joint1AxisExpVector, joint1AxisVector);
+  EXPECT_EQ(joint1AxisExpVector, model1Pose.Rot() * joint1AxisVector);
 
   const auto joint1Axis2 = joint1->Axis(1);
   ASSERT_NE(nullptr, joint1Axis2);
   Vector3d joint1Axis2Vector;
   EXPECT_TRUE(joint1Axis2->ResolveXyz(joint1Axis2Vector, "__model__").empty());
-  EXPECT_EQ(joint1Axis2ExpVector, joint1Axis2Vector);
+  EXPECT_EQ(joint1Axis2ExpVector, model1Pose.Rot() * joint1Axis2Vector);
 }
 
 // Function to remove any element in //joint/axis that is not <xyz>
@@ -526,9 +516,36 @@ static void removeNoneXyz(const sdf::ElementPtr &_elem)
     }
   }
 
-  for (auto el : toRemove)
+  for (const auto &el : toRemove)
   {
     _elem->RemoveChild(el);
+  }
+}
+
+void prepareModelForDirectComparison(sdf::ElementPtr _modelElem)
+{
+  if (_modelElem->HasElement("joint"))
+  {
+    for (auto joint = _modelElem->GetElement("joint"); joint;
+        joint = joint->GetNextElement("joint"))
+    {
+      if (joint->HasElement("axis"))
+      {
+        removeNoneXyz(joint->GetElement("axis"));
+      }
+      if (joint->HasElement("axis2"))
+      {
+        removeNoneXyz(joint->GetElement("axis2"));
+      }
+    }
+  }
+  if (_modelElem->HasElement("model"))
+  {
+    for (auto elem = _modelElem->GetElement("model"); elem;
+        elem = elem->GetNextElement("model"))
+    {
+      prepareModelForDirectComparison(elem);
+    }
   }
 }
 
@@ -546,21 +563,7 @@ void prepareForDirectComparison(sdf::ElementPtr _worldElem)
     }
     else
     {
-      if (elem->HasElement("joint"))
-      {
-        for (auto joint = elem->GetElement("joint"); joint;
-             joint = joint->GetNextElement("joint"))
-        {
-          if (joint->HasElement("axis"))
-          {
-            removeNoneXyz(joint->GetElement("axis"));
-          }
-          if (joint->HasElement("axis2"))
-          {
-            removeNoneXyz(joint->GetElement("axis2"));
-          }
-        }
-      }
+      prepareModelForDirectComparison(elem);
     }
   }
 
@@ -615,6 +618,82 @@ TEST(NestedModel, NestedModelWithFramesDirectComparison)
   std::stringstream expected;
   fs >> expected.rdbuf();
   EXPECT_EQ(expected.str(), sdfParsed->ToString());
+}
+
+//////////////////////////////////////////////////
+// Test DOM APIs with partially flattened model
+TEST(NestedModel, PartiallyFlattened)
+{
+  const std::string testFile =
+    sdf::filesystem::append(PROJECT_SOURCE_PATH, "test", "integration",
+      "partially_flattened.sdf");
+
+  // Load the SDF file
+  sdf::Root root;
+  auto errors = root.Load(testFile);
+  EXPECT_TRUE(errors.empty());
+
+  EXPECT_EQ(1u, root.WorldCount());
+  const sdf::World *world = root.WorldByIndex(0);
+  ASSERT_NE(nullptr, world);
+
+  // Get the outer model
+  const sdf::Model *outerModel = world->ModelByIndex(0);
+  ASSERT_NE(nullptr, outerModel);
+  EXPECT_EQ("ParentModel", outerModel->Name());
+  EXPECT_EQ(4u, outerModel->LinkCount());
+  EXPECT_NE(nullptr, outerModel->LinkByIndex(0));
+  EXPECT_NE(nullptr, outerModel->LinkByIndex(1));
+  EXPECT_NE(nullptr, outerModel->LinkByIndex(2));
+  EXPECT_NE(nullptr, outerModel->LinkByIndex(3));
+  EXPECT_EQ(nullptr, outerModel->LinkByIndex(4));
+
+  EXPECT_TRUE(outerModel->LinkNameExists("M1::L1"));
+  EXPECT_TRUE(outerModel->LinkNameExists("M1::L2"));
+  EXPECT_TRUE(outerModel->LinkNameExists("M1::L3"));
+  EXPECT_TRUE(outerModel->LinkNameExists("M1::L4"));
+
+  EXPECT_EQ(3u, outerModel->JointCount());
+  EXPECT_TRUE(outerModel->JointNameExists("M1::J1"));
+  EXPECT_TRUE(outerModel->JointNameExists("M1::J2"));
+  EXPECT_TRUE(outerModel->JointNameExists("M1::J3"));
+
+  EXPECT_EQ(3u, outerModel->FrameCount());
+  EXPECT_TRUE(outerModel->FrameNameExists("M1::__model__"));
+  EXPECT_TRUE(outerModel->FrameNameExists("M1::F1"));
+  EXPECT_TRUE(outerModel->FrameNameExists("M1::F2"));
+
+  EXPECT_EQ(1u, outerModel->ModelCount());
+  EXPECT_NE(nullptr, outerModel->ModelByIndex(0));
+  EXPECT_EQ(nullptr, outerModel->ModelByIndex(1));
+
+  // Get the middle model
+  const sdf::Model *midModel = outerModel->ModelByIndex(0);
+  ASSERT_NE(nullptr, midModel);
+  EXPECT_EQ("M1::M2", midModel->Name());
+  // TODO: the following expectations should be true, but ModelNameExists
+  // and ModelByName do not support names containing "::".
+  // EXPECT_TRUE(outerModel->ModelNameExists("M1::M2"));
+  // EXPECT_EQ(midModel, outerModel->ModelByName("M1::M2"));
+
+  EXPECT_EQ(1u, midModel->LinkCount());
+  EXPECT_NE(nullptr, midModel->LinkByIndex(0));
+  EXPECT_EQ(nullptr, midModel->LinkByIndex(1));
+
+  EXPECT_TRUE(midModel->LinkNameExists("L5"));
+
+  EXPECT_EQ(0u, midModel->JointCount());
+
+  EXPECT_EQ(0u, midModel->FrameCount());
+
+  EXPECT_EQ(0u, midModel->ModelCount());
+
+  // test nested names from outer model
+  // TODO: the following expectations also fail due to the limitations of
+  // ModelNameExists and ModelByName not supporting names containing "::".
+  // const std::string midLinkNestedName = "M1::M2::L5";
+  // EXPECT_TRUE(outerModel->LinkNameExists(midLinkNestedName));
+  // EXPECT_NE(nullptr, outerModel->LinkByName(midLinkNestedName));
 }
 
 //////////////////////////////////////////////////
@@ -691,6 +770,7 @@ TEST(NestedModel, NestedModelWithSiblingFrames)
     << "<sdf version='" << version << "'>"
     << "<world name='default'>"
     << "  <model name='ParentModel'>"
+    << "    <frame name='parentModelFrame'/>"
     << "    <frame name='testFrame'>"
     << "     <pose>" << testFramePose << "</pose>"
     << "    </frame>"
@@ -718,12 +798,13 @@ TEST(NestedModel, NestedModelWithSiblingFrames)
 
   const sdf::Model *parentModel = world->ModelByIndex(0);
   ASSERT_NE(nullptr, parentModel);
+  const sdf::Frame *parentModelFrame = parentModel->FrameByIndex(0);
+  ASSERT_NE(nullptr, parentModelFrame);
 
-  const sdf::Frame *nestedModel1Frame =
-      parentModel->FrameByName("M1::__model__");
-  ASSERT_NE(nullptr, nestedModel1Frame);
+  const sdf::Model *nestedModel1 = parentModel->ModelByName("M1");
+  ASSERT_NE(nullptr, nestedModel1);
 
-  EXPECT_EQ("testFrame", nestedModel1Frame->PoseRelativeTo());
+  EXPECT_EQ("testFrame", nestedModel1->PoseRelativeTo());
 
   using ignition::math::Pose3d;
   using ignition::math::Vector3d;
@@ -738,38 +819,40 @@ TEST(NestedModel, NestedModelWithSiblingFrames)
 
   Pose3d nestedModel1FramePose;
   EXPECT_TRUE(
-      nestedModel1Frame->SemanticPose().Resolve(nestedModel1FramePose).empty());
+      nestedModel1->SemanticPose().Resolve(nestedModel1FramePose).empty());
   EXPECT_EQ(nestedModel1FrameExpPose, nestedModel1FramePose);
+
+  auto parentSemPose = parentModel->SemanticPose();
 
   const auto *frame1 = parentModel->FrameByName("M1::F1");
   ASSERT_NE(nullptr, frame1);
   Pose3d frame1Pose;
-  EXPECT_TRUE(frame1->SemanticPose().Resolve(frame1Pose).empty());
-  EXPECT_EQ(frame1ExpPose, frame1Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(frame1Pose, "ParentModel::M1::F1").empty());
+  EXPECT_EQ(frame1ExpPose, frame1Pose.Inverse());
 
   const auto *frame2 = parentModel->FrameByName("M1::F2");
   ASSERT_NE(nullptr, frame2);
   Pose3d frame2Pose;
-  EXPECT_TRUE(frame2->SemanticPose().Resolve(frame2Pose).empty());
-  EXPECT_EQ(frame2ExpPose, frame2Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(frame2Pose, "ParentModel::M1::F2").empty());
+  EXPECT_EQ(frame2ExpPose, frame2Pose.Inverse());
 
   const auto *link1 = parentModel->LinkByName("M1::L1");
   ASSERT_NE(nullptr, link1);
   Pose3d link1Pose;
-  EXPECT_TRUE(link1->SemanticPose().Resolve(link1Pose).empty());
-  EXPECT_EQ(link1ExpPose, link1Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(link1Pose, "ParentModel::M1::L1").empty());
+  EXPECT_EQ(link1ExpPose, link1Pose.Inverse());
 
   const auto *link2 = parentModel->LinkByName("M1::L2");
   ASSERT_NE(nullptr, link2);
   Pose3d link2Pose;
-  EXPECT_TRUE(link2->SemanticPose().Resolve(link2Pose).empty());
-  EXPECT_EQ(link2ExpPose, link2Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(link2Pose, "ParentModel::M1::L2").empty());
+  EXPECT_EQ(link2ExpPose, link2Pose.Inverse());
 
   const auto *joint1 = parentModel->JointByName("M1::J1");
   ASSERT_NE(nullptr, joint1);
   Pose3d joint1Pose;
-  EXPECT_TRUE(joint1->SemanticPose().Resolve(joint1Pose, "__model__").empty());
-  EXPECT_EQ(joint1ExpPose, joint1Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(joint1Pose, "ParentModel::M1::J1").empty());
+  EXPECT_EQ(joint1ExpPose, joint1Pose.Inverse());
 }
 
 //////////////////////////////////////////////////
@@ -784,11 +867,12 @@ TEST(NestedModel, NestedFrameOnlyModel)
   const ignition::math::Pose3d model1Pose(10, 0, 0, 0, 0, IGN_PI/2);
 
   std::ostringstream stream;
-  std::string version = "1.7";
+  std::string version = "1.8";
   stream
     << "<sdf version='" << version << "'>"
     << "<world name='default'>"
     << "  <model name='ParentModel'>"
+    << "    <static>true</static>"
     << "    <include>"
     << "      <uri>" + MODEL_PATH + "</uri>"
     << "      <name>M1</name>"
@@ -804,9 +888,7 @@ TEST(NestedModel, NestedFrameOnlyModel)
 
   sdf::Root root;
   sdf::Errors errors = root.Load(sdfParsed);
-  for (auto e : errors)
-    std::cout << e.Message() << std::endl;
-  EXPECT_TRUE(errors.empty());
+  EXPECT_TRUE(errors.empty()) << errors;
 
   const sdf::World *world = root.WorldByIndex(0);
   ASSERT_NE(nullptr, world);
@@ -814,23 +896,21 @@ TEST(NestedModel, NestedFrameOnlyModel)
   const sdf::Model *parentModel = world->ModelByIndex(0);
   ASSERT_NE(nullptr, parentModel);
 
+  auto parentSemPose = parentModel->SemanticPose();
+
   using ignition::math::Pose3d;
   using ignition::math::Vector3d;
   // Expected poses for frames after nesting.
   Pose3d frame1ExpPose = model1Pose * Pose3d(0, 0, 0, IGN_PI, 0, 0);
   Pose3d frame2ExpPose = frame1ExpPose * Pose3d(0, 0, 0, 0, IGN_PI, 0);
 
-  const auto *frame1 = parentModel->FrameByName("M1::F1");
-  ASSERT_NE(nullptr, frame1);
   Pose3d frame1Pose;
-  EXPECT_TRUE(frame1->SemanticPose().Resolve(frame1Pose).empty());
-  EXPECT_EQ(frame1ExpPose, frame1Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(frame1Pose, "ParentModel::M1::F1").empty());
+  EXPECT_EQ(frame1ExpPose, frame1Pose.Inverse());
 
-  const auto *frame2 = parentModel->FrameByName("M1::F2");
-  ASSERT_NE(nullptr, frame2);
   Pose3d frame2Pose;
-  EXPECT_TRUE(frame2->SemanticPose().Resolve(frame2Pose).empty());
-  EXPECT_EQ(frame2ExpPose, frame2Pose);
+  EXPECT_TRUE(parentSemPose.Resolve(frame2Pose, "ParentModel::M1::F2").empty());
+  EXPECT_EQ(frame2ExpPose, frame2Pose.Inverse());
 }
 
 class PlacementFrame: public ::testing::Test
@@ -940,13 +1020,13 @@ class PlacementFrame: public ::testing::Test
 
     Pose3d testFramePose;
     {
-      sdf::Errors errors =
-          testFrame->SemanticPose().Resolve(testFramePose, "__model__");
-      EXPECT_TRUE(errors.empty()) << errors[0].Message();
+      sdf::Errors errors = parentModel->SemanticPose().Resolve(
+          testFramePose, parentModel->Name() + "::" + _testFrameName);
+      EXPECT_TRUE(errors.empty()) << errors;
     }
 
     // The expected pose of the test frame is precisely the placement pose
-    EXPECT_EQ(placementPose, testFramePose);
+    EXPECT_EQ(placementPose, testFramePose.Inverse());
   }
 
   protected: sdf::Root root;
@@ -1013,4 +1093,281 @@ TEST_F(PlacementFrame, ModelPlacementFrameAttribute)
       "model_with_placement_frame_and_pose_relative_to", "L4");
 }
 
-// TODO (addisu) Add NestedModelPlacementFrameAttribute tests
+//////////////////////////////////////////////////
+TEST(NestedReference, PoseRelativeTo)
+{
+  const std::string testFile =
+    sdf::filesystem::append(PROJECT_SOURCE_PATH, "test", "sdf",
+        "model_relative_to_nested_reference.sdf");
+
+  // Load the SDF file
+  sdf::Root root;
+  EXPECT_TRUE(root.Load(testFile).empty());
+
+  using Pose = ignition::math::Pose3d;
+  // Get the first model
+  const sdf::Model *model = root.ModelByIndex(0);
+  ASSERT_NE(nullptr, model);
+
+  auto parentSemPose = model->SemanticPose();
+  // The child models starting from M2 all reference an entity inside M1. Check
+  // that resolving poses works when referencing such entities.
+  {
+    Pose pose;
+    const sdf::Model *testModel = model->ModelByName("M2");
+    ASSERT_NE(nullptr, testModel);
+    EXPECT_TRUE(parentSemPose.Resolve(pose, "parent_model::M2").empty());
+    EXPECT_EQ(Pose(1, 0, -2, 0, IGN_PI_2, 0), pose.Inverse());
+  }
+  {
+    Pose pose;
+    const sdf::Model *testModel = model->ModelByName("M3");
+    ASSERT_NE(nullptr, testModel);
+    EXPECT_TRUE(parentSemPose.Resolve(pose, "parent_model::M3").empty());
+    EXPECT_EQ(Pose(1, 1, -3, 0, IGN_PI_2, 0), pose.Inverse());
+  }
+  {
+    Pose pose;
+    const sdf::Model *testModel = model->ModelByName("M4");
+    ASSERT_NE(nullptr, testModel);
+    EXPECT_TRUE(parentSemPose.Resolve(pose, "parent_model::M4").empty());
+    EXPECT_EQ(Pose(2, 0, -4, 0, IGN_PI_2, 0), pose.Inverse());
+  }
+  {
+    Pose pose;
+    const sdf::Model *testModel = model->ModelByName("M5");
+    ASSERT_NE(nullptr, testModel);
+    EXPECT_TRUE(parentSemPose.Resolve(pose, "parent_model::M5").empty());
+    EXPECT_EQ(Pose(2, 0, -6, 0, IGN_PI_2, 0), pose.Inverse());
+  }
+  {
+    Pose pose;
+    const sdf::Model *testModel = model->ModelByName("M6");
+    ASSERT_NE(nullptr, testModel);
+    EXPECT_TRUE(parentSemPose.Resolve(pose, "parent_model::M6").empty());
+    EXPECT_EQ(Pose(1, 1, -6, IGN_PI_2, IGN_PI_2, 0), pose.Inverse());
+  }
+  {
+    Pose pose;
+    const sdf::Model *testModel = model->ModelByName("M7");
+    ASSERT_NE(nullptr, testModel);
+    EXPECT_TRUE(parentSemPose.Resolve(pose, "parent_model::M7").empty());
+    EXPECT_EQ(Pose(1, -6, -1, 0, 0, -IGN_PI_2), pose.Inverse());
+  }
+}
+
+//////////////////////////////////////////////////
+TEST(NestedReference, PoseRelativeToInWorld)
+{
+  const std::string testFile =
+    sdf::filesystem::append(PROJECT_SOURCE_PATH, "test", "sdf",
+        "world_relative_to_nested_reference.sdf");
+
+  // Load the SDF file
+  sdf::Root root;
+  EXPECT_TRUE(root.Load(testFile).empty());
+  const sdf::World *world = root.WorldByIndex(0);
+  ASSERT_NE(nullptr, world);
+
+  using Pose = ignition::math::Pose3d;
+
+  // The //world/frame elements reference an entity inside M1. Check
+  // that resolving poses works when referencing such entities.
+  {
+    Pose pose;
+    const sdf::Frame *testFrame = world->FrameByName("F2");
+    ASSERT_NE(nullptr, testFrame);
+    EXPECT_TRUE(testFrame->SemanticPose().Resolve(pose).empty());
+    EXPECT_EQ(Pose(1, 0, -2, 0, IGN_PI_2, 0), pose);
+  }
+  {
+    Pose pose;
+    const sdf::Frame *testFrame = world->FrameByName("F3");
+    ASSERT_NE(nullptr, testFrame);
+    EXPECT_TRUE(testFrame->SemanticPose().Resolve(pose).empty());
+    EXPECT_EQ(Pose(1, 1, -3, 0, IGN_PI_2, 0), pose);
+  }
+  {
+    Pose pose;
+    const sdf::Frame *testFrame = world->FrameByName("F4");
+    ASSERT_NE(nullptr, testFrame);
+    EXPECT_TRUE(testFrame->SemanticPose().Resolve(pose).empty());
+    EXPECT_EQ(Pose(2, 0, -4, 0, IGN_PI_2, 0), pose);
+  }
+  {
+    Pose pose;
+    const sdf::Frame *testFrame = world->FrameByName("F5");
+    ASSERT_NE(nullptr, testFrame);
+    EXPECT_TRUE(testFrame->SemanticPose().Resolve(pose).empty());
+    EXPECT_EQ(Pose(2, 0, -6, 0, IGN_PI_2, 0), pose);
+  }
+  {
+    Pose pose;
+    const sdf::Frame *testFrame = world->FrameByName("F6");
+    ASSERT_NE(nullptr, testFrame);
+    EXPECT_TRUE(testFrame->SemanticPose().Resolve(pose).empty());
+    EXPECT_EQ(Pose(1, 1, -6, IGN_PI_2, IGN_PI_2, 0), pose);
+  }
+  {
+    Pose pose;
+    const sdf::Frame *testFrame = world->FrameByName("F7");
+    ASSERT_NE(nullptr, testFrame);
+    EXPECT_TRUE(testFrame->SemanticPose().Resolve(pose).empty());
+    EXPECT_EQ(Pose(1, -6, -1, 0, 0, -IGN_PI_2), pose);
+  }
+}
+
+/////////////////////////////////////////////////
+TEST(NestedReference, PlacementFrameAttribute)
+{
+  const std::string testFile =
+      sdf::filesystem::append(PROJECT_SOURCE_PATH, "test", "integration",
+          "model", "model_with_nested_placement_frame_attribute", "model.sdf");
+
+  // Load the SDF file
+  sdf::Root root;
+  sdf::Errors errors = root.Load(testFile);
+  EXPECT_TRUE(errors.empty()) << errors;
+
+  auto *model = root.ModelByIndex(0);
+  ASSERT_NE(nullptr, model);
+
+  ignition::math::Pose3d pose;
+  errors = model->SemanticPose().Resolve(pose);
+  EXPECT_TRUE(errors.empty()) << errors;
+  EXPECT_EQ(ignition::math::Pose3d(0, -2, 10, 0, 0, 0), pose);
+}
+
+/////////////////////////////////////////////////
+TEST(NestedReference, PlacementFrameElement)
+{
+  const std::string modelRootPath = sdf::filesystem::append(
+      PROJECT_SOURCE_PATH, "test", "integration", "model");
+
+  sdf::setFindCallback([&](const std::string &_file)
+  {
+    return sdf::filesystem::append(modelRootPath, _file);
+  });
+
+  // Test with //world/include without overriding the placement_frame
+  {
+    std::ostringstream stream;
+    stream << R"(
+    <sdf version='1.8'>
+      <world name='default'>
+        <include>
+          <uri>model_with_nested_placement_frame_attribute</uri>
+        </include>
+      </world>
+    </sdf>)";
+    // Load the SDF file
+    sdf::Root root;
+    sdf::Errors errors = root.LoadSdfString(stream.str());
+    EXPECT_TRUE(errors.empty()) << errors;
+
+    auto *world = root.WorldByIndex(0);
+    ASSERT_NE(nullptr, world);
+    auto *model = world->ModelByIndex(0);
+    ASSERT_NE(nullptr, model);
+
+    ignition::math::Pose3d pose;
+    errors = model->SemanticPose().Resolve(pose);
+    EXPECT_TRUE(errors.empty()) << errors;
+    EXPECT_EQ(ignition::math::Pose3d(0, -2, 10, 0, 0, 0), pose);
+  }
+
+  // Test with //world/include overriding the placement_frame
+  {
+    const ignition::math::Pose3d placementPose(0, 10, 0, IGN_PI_2, 0, 0);
+    std::ostringstream stream;
+    stream << R"(
+    <sdf version='1.8'>
+      <world name='default'>
+        <frame name="world_frame"/>
+        <include>
+          <uri>model_with_nested_placement_frame_attribute</uri>
+          <placement_frame>child_model::link3</placement_frame>
+          <pose>)" << placementPose << R"(</pose>"
+        </include>
+      </world>
+    </sdf>)";
+    // Load the SDF file
+    sdf::Root root;
+    sdf::Errors errors = root.LoadSdfString(stream.str());
+    EXPECT_TRUE(errors.empty()) << errors;
+
+    auto *world = root.WorldByIndex(0);
+    ASSERT_NE(nullptr, world);
+
+    auto *world_frame = world->FrameByName("world_frame");
+    ASSERT_NE(nullptr, world_frame);
+    ignition::math::Pose3d pose;
+    errors = world_frame->SemanticPose().Resolve(
+        pose, "model_placement_frame::child_model::link3");
+    EXPECT_TRUE(errors.empty()) << errors;
+    EXPECT_EQ(placementPose, pose.Inverse());
+  }
+
+
+  // Test with //model/include without overriding the placement_frame
+  {
+    std::ostringstream stream;
+    stream << R"(
+    <sdf version='1.8'>
+      <model name='parent_model'>
+        <include>
+          <uri>model_with_nested_placement_frame_attribute</uri>
+        </include>
+      </model>
+    </sdf>)";
+    // Load the SDF file
+    sdf::Root root;
+    sdf::Errors errors = root.LoadSdfString(stream.str());
+    EXPECT_TRUE(errors.empty()) << errors;
+
+    auto *parentModel = root.ModelByIndex(0);
+    ASSERT_NE(nullptr, parentModel);
+    auto *model = parentModel->ModelByIndex(0);
+    ASSERT_NE(nullptr, model);
+
+    ignition::math::Pose3d pose;
+    errors = model->SemanticPose().Resolve(pose);
+    EXPECT_TRUE(errors.empty()) << errors;
+    EXPECT_EQ(ignition::math::Pose3d(0, -2, 10, 0, 0, 0), pose);
+  }
+
+  // Test with //model/include overriding the placement_frame
+  {
+    const ignition::math::Pose3d placementPose(0, 10, 0, IGN_PI_2, 0, 0);
+    std::ostringstream stream;
+    stream << R"(
+    <sdf version='1.8'>
+      <model name='parent_model'>
+        <frame name="parent_model_frame"/>
+        <include>
+          <uri>model_with_nested_placement_frame_attribute</uri>
+          <placement_frame>child_model::link3</placement_frame>
+          <pose>)" << placementPose << R"(</pose>"
+        </include>
+      </model>
+    </sdf>)";
+    // Load the SDF file
+    sdf::Root root;
+    sdf::Errors errors = root.LoadSdfString(stream.str());
+    EXPECT_TRUE(errors.empty()) << errors;
+
+    auto *parentModel = root.ModelByIndex(0);
+    ASSERT_NE(nullptr, parentModel);
+    auto *model = parentModel ->ModelByIndex(0);
+    ASSERT_NE(nullptr, model);
+
+    auto *parentModelFrame = parentModel->FrameByName("parent_model_frame");
+    auto *link = model->LinkByName("child_model::link3");
+    ASSERT_NE(nullptr, link);
+    ignition::math::Pose3d pose;
+    errors = parentModelFrame->SemanticPose().Resolve(
+        pose, "model_placement_frame::child_model::link3");
+    EXPECT_TRUE(errors.empty()) << errors;
+    EXPECT_EQ(placementPose, pose.Inverse());
+  }
+}